Move TypeSystemSwiftTypeRef into its own file. (NFC)
